🚚 Rename: userMe.types.ts 폴더 변경
