+++ src/features/users/apis/getCurrentUser.ts
@@ -1,4 +1,4 @@
 import { get } from '@/shared/apis/fetchInstance';
-import type { UserMe } from '@/shared/types/userMe.types';
+import type { UserMe } from '@/features/users/apis/userMe.types';
 
 export const getCurrentUser = () => get<UserMe>('users/me');
